Refactor report document assembly seams
