+++ apps/server/vibesensor/use_cases/history/report_document/composition.py
@@ -2,52 +2,30 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable
 from dataclasses import replace
 
-from vibesensor.report_i18n import normalize_lang
-from vibesensor.report_i18n import tr as _tr
-from vibesensor.shared.boundaries.reporting import PreparedReportFacts, PreparedReportInput
+from vibesensor.shared.boundaries.reporting import PreparedReportInput
 from vibesensor.shared.boundaries.reporting.document import (
     AppendixAData,
     NextStep,
     ReportDocument,
+    VerdictPageData,
 )
-from vibesensor.shared.boundaries.reporting.facts import ReportRunFacts
-from vibesensor.shared.report_presentation import (
-    coverage_label,
-    coverage_notes,
-    proof_caveat_text,
-    runner_up_corner,
-)
-from vibesensor.shared.time_utils import (
-    format_timestamp_in_recorded_timezone,
-    format_utc_timestamp,
-    utc_now_iso,
-)
-from vibesensor.shared.types.json_types import JsonValue
+from vibesensor.shared.time_utils import format_utc_timestamp
 
-from ._candidate_resolver import PrimaryCandidateContext, resolve_primary_report_candidate
 from ._card_builder import build_system_cards
+from .appendix_c import build_appendix_c_data
+from .document_context import ReportDocumentContext, build_report_document_context
 from .location_appendix import build_appendix_b_data
 from .measurements import _measurement_rows
 from .narrative_summaries import _proof_summary_text
 from .pattern_evidence import build_pattern_evidence
 from .peak_table import build_peak_rows
 from .report_sections import build_data_trust, build_next_steps
-from .section_context import (
-    AppendixAContext,
-    AppendixBContext,
-    AppendixCContext,
-    VerdictPageContext,
-)
-from .sections import _build_appendix_c_data, _build_timeline_graph_data, _build_traceability_rows
+from .timeline_graph import build_timeline_graph_data
+from .traceability import build_traceability_rows
 from .verdict_page import build_observed_signature, build_verdict_page_data
-from .workflow_appendix import (
-    build_appendix_a_data,
-    build_ranked_candidates,
-    build_recapture_assessment,
-)
+from .workflow_appendix import build_appendix_a_data
 
 __all__ = ["compose_report_document"]
 
@@ -55,243 +33,155 @@ __all__ = ["compose_report_document"]
 def compose_report_document(prepared: PreparedReportInput) -> ReportDocument:
     """Compose the canonical report document from prepared report input."""
 
-    lang = str(normalize_lang(prepared.language))
-
-    def tr(key: str, **kw: JsonValue) -> str:
-        return str(_tr(lang, key, **kw))
-
-    test_run = prepared.domain_test_run
-    report_facts = prepared.report_facts
-    run_facts = report_facts.run
-    sensor_facts = report_facts.sensor
-    decision_facts = report_facts.decision
-    prepared_findings = report_facts.findings
-    coverage = sensor_facts.coverage
-    active_locations = tuple(coverage.active_locations)
-    coverage_label_text = coverage_label(
-        expected_locations=coverage.expected_locations,
-        active_locations=coverage.active_locations,
-        missing_locations=coverage.missing_locations,
-        partial_locations=coverage.partial_locations,
-        tr=tr,
-    )
-    coverage_notes_items = tuple(
-        coverage_notes(
-            missing_locations=coverage.missing_locations,
-            partial_locations=coverage.partial_locations,
-            tr=tr,
-        )
-    )
-    proof_caveat = proof_caveat_text(
-        primary_candidate_facts=decision_facts.primary_candidate,
-        action_status_key=decision_facts.action_status_key,
-        location_confidence_key=decision_facts.location_confidence_key,
-        tr=tr,
-    )
-    runner_up = runner_up_corner(sensor_facts.active_intensity, tr=tr)
-    speed_window_label = str(decision_facts.primary_candidate.primary_speed or "").strip() or None
-    recapture = build_recapture_assessment(
-        aggregate=test_run,
-        primary_candidate_facts=decision_facts.primary_candidate,
-        location_confidence_key=decision_facts.location_confidence_key,
-        expected_locations=coverage.expected_locations,
-        active_locations=coverage.active_locations,
-        suitability_checks=decision_facts.suitability_checks,
-        warnings=decision_facts.warnings,
-        lang=lang,
-        tr=tr,
-    )
-    verdict_context = VerdictPageContext(
-        action_status_key=decision_facts.action_status_key,
-        location_confidence_key=decision_facts.location_confidence_key,
-        alternative_source_visible=decision_facts.alternative_source_visible,
-        active_locations=active_locations,
-        coverage_label=coverage_label_text,
-        proof_caveat=proof_caveat,
-        runner_up_corner=runner_up,
-        speed_window_label=speed_window_label,
-        recapture=recapture,
-    )
-    appendix_a_context = AppendixAContext(
-        action_status_key=decision_facts.action_status_key,
-        alternative_source_visible=decision_facts.alternative_source_visible,
-        ranked_candidates=build_ranked_candidates(test_run, tr=tr),
-        recapture=recapture,
-    )
-    appendix_b_context = AppendixBContext(
-        action_status_key=decision_facts.action_status_key,
-        location_confidence_key=decision_facts.location_confidence_key,
-        active_locations=active_locations,
-        coverage_label=coverage_label_text,
-        coverage_notes=coverage_notes_items,
-        runner_up_corner=runner_up,
-    )
-    appendix_c_context = AppendixCContext(
-        speed_window_label=speed_window_label,
-        proof_caveat=proof_caveat,
-    )
-    verdict_page = build_verdict_page_data(
-        aggregate=test_run,
-        primary_candidate_facts=decision_facts.primary_candidate,
-        duration_text=run_facts.duration_text,
-        verdict_context=verdict_context,
-        suitability_checks=decision_facts.suitability_checks,
-        warnings=decision_facts.warnings,
-        lang=lang,
-        tr=tr,
-    )
+    context = build_report_document_context(prepared)
     appendix_a = build_appendix_a_data(
-        aggregate=test_run,
-        appendix_context=appendix_a_context,
-        tr=tr,
+        aggregate=context.test_run,
+        appendix_context=context.appendix_a_context,
+        tr=context.tr,
     )
     appendix_b = build_appendix_b_data(
-        aggregate=test_run,
-        primary_candidate_facts=decision_facts.primary_candidate,
-        active_sensor_intensity=sensor_facts.active_intensity,
-        appendix_context=appendix_b_context,
-        tr=tr,
-    )
-    primary = resolve_primary_report_candidate(
-        aggregate=test_run,
-        facts=decision_facts.primary_candidate,
-        tr=tr,
-        lang=lang,
+        aggregate=context.test_run,
+        primary_candidate_facts=context.decision_facts.primary_candidate,
+        active_sensor_intensity=context.sensor_facts.active_intensity,
+        appendix_context=context.appendix_b_context,
+        tr=context.tr,
     )
     data_trust = build_data_trust(
-        suitability_checks=decision_facts.suitability_checks,
-        warnings=decision_facts.warnings,
-        lang=lang,
-        tr=tr,
+        suitability_checks=context.decision_facts.suitability_checks,
+        warnings=context.decision_facts.warnings,
+        lang=context.lang,
+        tr=context.tr,
     )
-    proof_summary = _proof_summary_text(
-        test_run,
-        primary,
-        report_facts,
-        runner_up_corner=verdict_context.runner_up_corner,
-        tr=tr,
-    )
-    run_datetime = _report_date_text(run_facts)
-    findings = list(prepared_findings.all_findings)
-    verdict_page = replace(
-        verdict_page,
-        proof_summary=proof_summary,
-        timeline_graph=_build_timeline_graph_data(
-            report_facts,
-            duration_s=run_facts.duration_s,
-        ),
-    )
+    verdict_page = _build_verdict_page(context=context)
     return ReportDocument(
-        title=tr("REPORT_FOOTER_TITLE"),
-        run_datetime=run_datetime,
-        run_id=run_facts.run_id,
-        duration_text=run_facts.duration_text,
-        start_time_utc=format_utc_timestamp(run_facts.start_time_utc),
-        end_time_utc=format_utc_timestamp(run_facts.end_time_utc),
-        sample_rate_hz=run_facts.sample_rate_hz,
-        tire_spec_text=run_facts.tire_spec_text,
-        sample_count=run_facts.sample_count,
-        sensor_count=primary.sensor_count,
-        sensor_locations=list(sensor_facts.active_locations),
-        sensor_model=run_facts.sensor_model,
-        firmware_version=run_facts.firmware_version,
-        car_name=run_facts.car_name,
-        car_type=run_facts.car_type,
-        observed=build_observed_signature(primary, tr=tr),
+        title=context.tr("REPORT_FOOTER_TITLE"),
+        run_datetime=context.run_datetime,
+        run_id=context.run_facts.run_id,
+        duration_text=context.run_facts.duration_text,
+        start_time_utc=format_utc_timestamp(context.run_facts.start_time_utc),
+        end_time_utc=format_utc_timestamp(context.run_facts.end_time_utc),
+        sample_rate_hz=context.run_facts.sample_rate_hz,
+        tire_spec_text=context.run_facts.tire_spec_text,
+        sample_count=context.run_facts.sample_count,
+        sensor_count=context.primary.sensor_count,
+        sensor_locations=list(context.sensor_facts.active_locations),
+        sensor_model=context.run_facts.sensor_model,
+        firmware_version=context.run_facts.firmware_version,
+        car_name=context.run_facts.car_name,
+        car_type=context.run_facts.car_type,
+        observed=build_observed_signature(context.primary, tr=context.tr),
         system_cards=build_system_cards(
-            test_run,
-            primary,
-            lang,
-            tr,
+            context.test_run,
+            context.primary,
+            context.lang,
+            context.tr,
         ),
         next_steps=list(
             _resolve_next_steps(
-                primary=primary,
-                report_facts=report_facts,
+                context=context,
                 appendix_a=appendix_a,
-                lang=lang,
-                tr=tr,
             )
         ),
         data_trust=data_trust,
         pattern_evidence=build_pattern_evidence(
-            aggregate=test_run,
-            origin=run_facts.origin,
-            primary=primary,
-            lang=lang,
-            tr=tr,
+            aggregate=context.test_run,
+            origin=context.run_facts.origin,
+            primary=context.primary,
+            lang=context.lang,
+            tr=context.tr,
         ),
         peak_rows=build_peak_rows(
-            run_facts.peak_table_rows,
-            findings=findings,
-            lang=lang,
-            tr=tr,
+            context.run_facts.peak_table_rows,
+            findings=list(context.findings),
+            lang=context.lang,
+            tr=context.tr,
         ),
-        lang=lang,
-        certainty_tier_key=primary.tier,
-        findings=findings,
-        top_causes=list(prepared_findings.top_causes),
-        sensor_intensity_by_location=list(sensor_facts.active_intensity),
-        location_hotspot_rows=list(sensor_facts.location_hotspot_rows),
+        lang=context.lang,
+        certainty_tier_key=context.primary.tier,
+        findings=list(context.findings),
+        top_causes=list(context.top_causes),
+        sensor_intensity_by_location=list(context.sensor_facts.active_intensity),
+        location_hotspot_rows=list(context.sensor_facts.location_hotspot_rows),
         verdict_page=verdict_page,
         appendix_a=appendix_a,
         appendix_b=appendix_b,
-        appendix_c=_build_appendix_c_data(
-            primary=primary,
-            aggregate=test_run,
+        appendix_c=build_appendix_c_data(
+            primary=context.primary,
+            aggregate=context.test_run,
             measurements=_measurement_rows(
-                run_facts,
-                aggregate=test_run,
-                tr=tr,
+                context.run_facts,
+                aggregate=context.test_run,
+                tr=context.tr,
             ),
-            report_facts=report_facts,
-            appendix_context=appendix_c_context,
+            report_facts=context.report_facts,
+            appendix_context=context.appendix_c_context,
             data_trust=list(data_trust),
-            tr=tr,
+            tr=context.tr,
         ),
         traceability_rows=list(
-            _build_traceability_rows(
-                date_str=run_datetime,
-                run_id=run_facts.run_id,
-                tire_spec_text=run_facts.tire_spec_text,
-                sensor_model=run_facts.sensor_model,
-                firmware_version=run_facts.firmware_version,
-                sample_count=run_facts.sample_count,
-                sample_rate_hz=run_facts.sample_rate_hz,
-                tr=tr,
+            build_traceability_rows(
+                date_str=context.run_datetime,
+                run_id=context.run_facts.run_id,
+                tire_spec_text=context.run_facts.tire_spec_text,
+                sensor_model=context.run_facts.sensor_model,
+                firmware_version=context.run_facts.firmware_version,
+                sample_count=context.run_facts.sample_count,
+                sample_rate_hz=context.run_facts.sample_rate_hz,
+                tr=context.tr,
             )
+        ),
+    )
+
+
+def _build_verdict_page(
+    *,
+    context: ReportDocumentContext,
+) -> VerdictPageData:
+    verdict_page = build_verdict_page_data(
+        aggregate=context.test_run,
+        primary_candidate_facts=context.decision_facts.primary_candidate,
+        duration_text=context.run_facts.duration_text,
+        verdict_context=context.verdict_page_context,
+        suitability_checks=context.decision_facts.suitability_checks,
+        warnings=context.decision_facts.warnings,
+        lang=context.lang,
+        tr=context.tr,
+    )
+    proof_summary = _proof_summary_text(
+        context.test_run,
+        context.primary,
+        context.report_facts,
+        runner_up_corner=context.verdict_page_context.runner_up_corner,
+        tr=context.tr,
+    )
+    return replace(
+        verdict_page,
+        proof_summary=proof_summary,
+        timeline_graph=build_timeline_graph_data(
+            context.report_facts,
+            duration_s=context.run_facts.duration_s,
         ),
     )
 
 
 def _resolve_next_steps(
     *,
-    primary: PrimaryCandidateContext,
-    report_facts: PreparedReportFacts,
+    context: ReportDocumentContext,
     appendix_a: AppendixAData,
-    lang: str,
-    tr: Callable[..., str],
 ) -> tuple[NextStep, ...]:
-    recapture_mode = report_facts.decision.action_status_key == "recapture_before_acting"
+    recapture_mode = context.decision_facts.action_status_key == "recapture_before_acting"
     if recapture_mode:
         return tuple(NextStep(action=action) for action in appendix_a.capture_changes)
     return tuple(
         build_next_steps(
-            recommended_actions=report_facts.decision.recommended_actions,
-            primary_source=primary.primary_source,
-            primary_location=primary.primary_location,
-            tier=primary.tier,
-            cert_reason=primary.certainty_reason or tr("REPORT_CAPTURE_ISSUE_GENERIC"),
+            recommended_actions=context.decision_facts.recommended_actions,
+            primary_source=context.primary.primary_source,
+            primary_location=context.primary.primary_location,
+            tier=context.primary.tier,
+            cert_reason=context.primary.certainty_reason
+            or context.tr("REPORT_CAPTURE_ISSUE_GENERIC"),
             recapture_mode=recapture_mode,
-            lang=lang,
-            tr=tr,
+            lang=context.lang,
+            tr=context.tr,
         )
     )
-
-
-def _report_date_text(run_facts: ReportRunFacts) -> str:
-    report_date = run_facts.report_date or utc_now_iso()
-    return format_timestamp_in_recorded_timezone(
-        report_date,
-        run_facts.recorded_utc_offset_seconds,
-    ) or str(report_date)
